db code completed

+++ db/migration/000001_init_schema.up.sql
@@ -14,9 +14,9 @@ CREATE TABLE "entries" (
 );
 
 CREATE TABLE "transfers" (
-                             "id" int PRIMARY KEY,
-                             "from_account_id" int NOT NULL,
-                             "to_account_id" int NOT NULL,
+                             "id" bigserial PRIMARY KEY,
+                             "from_account_id" bigint NOT NULL,
+                             "to_account_id" bigint NOT NULL,
                              "amount" bigint NOT NULL,
                              "created_at" timestamptz NOT NULL DEFAULT (now())
 );
